+++ NEWS.md
@@ -1,3 +1,12 @@
+# RWDataPlyr 0.6.6.9000 (development version)
+
+* Added `bigrdf_to_rwtbl()` and `bigrdf_save()`
+* "Scale" is now a valid column when creating rwtbl in `rdf_to_rwtbl()`, `rdf_aggregate()`, and `rwscen_aggregate()` (#117)
+* Fixed error in RWDataPlyr Workflow vignette; now correctly references `rwd_agg_template()` (#115)
+* Fixed typo in `rw_scen_aggregate()` and `rdf_to_rwtbl()` documentation. "Unit" is a valid column; "Units" is not valid. (#114)
+* Updated call to `xts::period.apply()` to specify `FUN = colMeans` per updates to xts (#118)
+* Added new 'between' eval operators to `rwd_agg`s. (#113)
+
 # RWDataPlyr 0.6.6
 
 *Release January 12, 2026*